requirements.txt
alembic==1.8.0
asttokens==2.0.5
attrs==21.4.0
autopage==0.5.1
backcall==0.2.0
Bootstrap-Flask==2.0.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorlog==6.6.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dominate==2.6.0
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
fonttools==4.33.3
greenlet==1.1.2
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.3
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
optuna==2.10.1
packaging==21.3
pandas==1.4.2
parso==0.8.3
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prettytable==3.3.0
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.37
stack-data==0.3.0
stevedore==3.5.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
urllib3==1.26.9
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
xgboost==1.6.1
